requirements.txt
piexif==1.1.3
Pillow==5.0.0
